resource manager rework wip
