+++ project/carpg.vcxproj
@@ -143,9 +143,10 @@
     <ClInclude Include="source\core\Timer.h" />
     <ClInclude Include="source\core\Tokenizer.h" />
     <ClInclude Include="source\core\VertexData.h" />
-    <ClInclude Include="source\engine\Animesh.h" />
+    <ClInclude Include="source\engine\Mesh.h" />
     <ClInclude Include="source\engine\Engine.h" />
     <ClInclude Include="source\engine\KeyStates.h" />
+    <ClInclude Include="source\engine\MeshInstance.h" />
     <ClInclude Include="source\engine\ParticleSystem.h" />
     <ClInclude Include="source\engine\Physics.h" />
     <ClInclude Include="source\engine\Resource.h" />
@@ -341,9 +342,10 @@
     <ClCompile Include="source\core\Text.cpp" />
     <ClCompile Include="source\core\Timer.cpp" />
     <ClCompile Include="source\core\Tokenizer.cpp" />
-    <ClCompile Include="source\engine\Animesh.cpp" />
+    <ClCompile Include="source\engine\Mesh.cpp" />
     <ClCompile Include="source\engine\Engine.cpp" />
     <ClCompile Include="source\engine\KeyStates.cpp" />
+    <ClCompile Include="source\engine\MeshInstance.cpp" />
     <ClCompile Include="source\engine\ParticleSystem.cpp" />
     <ClCompile Include="source\engine\Physics.cpp" />
     <ClCompile Include="source\engine\ResourceManager.cpp" />
@@ -528,7 +530,6 @@
     <Text Include="..\doc\readme.txt" />
     <Text Include="..\doc\readme_eng.txt" />
     <Text Include="..\doc\system.txt" />
-    <Text Include="..\doc\todo last.txt" />
     <Text Include="..\doc\todo.txt" />
   </ItemGroup>
   <ItemGroup>
