+++ project/carpg.vcxproj.filters
@@ -579,9 +579,6 @@
     <ClCompile Include="source\gui\Font.cpp">
       <Filter>gui</Filter>
     </ClCompile>
-    <ClCompile Include="source\engine\Animesh.cpp">
-      <Filter>2. engine</Filter>
-    </ClCompile>
     <ClCompile Include="source\engine\Engine.cpp">
       <Filter>2. engine</Filter>
     </ClCompile>
@@ -608,6 +605,12 @@
     </ClCompile>
     <ClCompile Include="source\game\stats\Buff.cpp">
       <Filter>3. game\stats</Filter>
+    </ClCompile>
+    <ClCompile Include="source\engine\Mesh.cpp">
+      <Filter>2. engine</Filter>
+    </ClCompile>
+    <ClCompile Include="source\engine\MeshInstance.cpp">
+      <Filter>2. engine</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1166,9 +1169,6 @@
     <ClInclude Include="source\gui\Font.h">
       <Filter>gui</Filter>
     </ClInclude>
-    <ClInclude Include="source\engine\Animesh.h">
-      <Filter>2. engine</Filter>
-    </ClInclude>
     <ClInclude Include="source\engine\Engine.h">
       <Filter>2. engine</Filter>
     </ClInclude>
@@ -1211,6 +1211,12 @@
     <ClInclude Include="source\game\stats\Buff.h">
       <Filter>3. game\stats</Filter>
     </ClInclude>
+    <ClInclude Include="source\engine\Mesh.h">
+      <Filter>2. engine</Filter>
+    </ClInclude>
+    <ClInclude Include="source\engine\MeshInstance.h">
+      <Filter>2. engine</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="source\resource\resource.rc">
@@ -1237,9 +1243,6 @@
       <Filter>0. doc</Filter>
     </Text>
     <Text Include="..\doc\todo.txt">
-      <Filter>0. doc</Filter>
-    </Text>
-    <Text Include="..\doc\todo last.txt">
       <Filter>0. doc</Filter>
     </Text>
   </ItemGroup>
