+++ project/source/core/Core.h
@@ -86,8 +86,8 @@ typedef const char* cstring;
 #ifndef COMMON_ONLY
 typedef FMOD::Sound* SOUND;
 #endif
-struct Animesh;
-struct AnimeshInstance;
+struct Mesh;
+struct MeshInstance;
 
 //-----------------------------------------------------------------------------
 // Delegates
