+++ project/source/game/entity/Bullet.cpp
@@ -4,7 +4,7 @@
 #include "Bullet.h"
 #include "SaveState.h"
 #include "ResourceManager.h"
-#include "Animesh.h"
+#include "Mesh.h"
 #include "Unit.h"
 #include "Spell.h"
 #include "ParticleSystem.h"
