+++ project/source/game/entity/Chest.cpp
@@ -38,7 +38,7 @@ void Chest::Save(HANDLE file, bool local)
 
 	if(local)
 	{
-		AnimeshInstance::Group& group = ani->groups[0];
+		MeshInstance::Group& group = ani->groups[0];
 		if(group.IsPlaying())
 		{
 			WriteFile(file, &group.state, sizeof(group.state), &tmp, nullptr);
@@ -113,11 +113,11 @@ void Chest::Load(HANDLE file, bool local)
 		int b;
 		ReadFile(file, &b, sizeof(b), &tmp, nullptr);
 
-		ani = new AnimeshInstance(Game::Get().aSkrzynia);
+		ani = new MeshInstance(Game::Get().aSkrzynia);
 
 		if(b != 0)
 		{
-			AnimeshInstance::Group& group = ani->groups[0];
+			MeshInstance::Group& group = ani->groups[0];
 			group.anim = &ani->ani->anims[0];
 			group.state = b;
 			ReadFile(file, &group.time, sizeof(group.time), &tmp, nullptr);
